# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    This API is for the WASAText Project
    a full stack messaging application
  version: 0.0.1
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
paths:
...
  /conversations/{chatId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
      - $ref: "#/components/parameters/MessageID"

    delete:
      summary: Delete a message
      description: Delete a message in a Conversation

      tags: ["Messages"]

      operationId: deleteMessage

      responses:
        "204":
          description: Message deleted successfuly
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
...
  /messages/{messageId}/forward:
    parameters:
      - $ref: "#/components/parameters/MessageID"

    post:
      summary: Forward a message
      description: Forwards a message to a conversation

      tags: ["Messages"]

      operationId: forwardMessage

      requestBody:
        description: |
          The Conversation ID on which 
          the message will be forwarded

        content:
          application/json:
            schema:
              description: Conversation ID and Message ID
              type: object
              properties:
                conversationID:
                  $ref: "#/components/schemas/UUID"

                messageID:
                  $ref: "#/components/schemas/UUID"

      responses:
        "204":
          description: Message forwarded successfuly

        "401":
          description: Unauthorized

        "404":
          description: Message not found

        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    UUID:
      description: "Universally Unique Identifier"
      type: string
      format: uuid
      minLength: 1
      maxLength: 32
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ConversationID:
      name: chatId
      in: path
      required: true
      description: Conversation Identifier
      schema:
        $ref: "#/components/schemas/UUID"
    MessageID:
      name: messageId
      in: path
      required: true
      description: Message Identifier
      schema:
        $ref: "#/components/schemas/UUID"
...